LPD-6742 Use DB Partition Migration Validator since it involves the migration but only validates, does not migrate. Also, for consistency with other tools.
