+++ modules/util/portal-tools-db-partition-migration-validator/source-formatter-suppressions.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+
+<suppressions>
+	<checkstyle>
+		<suppress checks="CompanyIterationCheck" files="src/main/java/com/liferay/portal/tools/db/partition/migration/validator/util/DatabaseUtil\.java" />
+		<suppress checks="CompanyIterationCheck" files="src/main/java/com/liferay/portal/tools/db/partition/migration/validator/util/Validator\.java" />
+		<suppress checks="CompanyIterationCheck" files="src/test/java/com/liferay/portal/tools/db/partition/migration/validator/DBPartitionMigrationValidatorExportTest\.java" />
+		<suppress checks="CompanyIterationCheck" files="src/test/java/com/liferay/portal/tools/db/partition/migration/validator/util/DatabaseMockupUtil\.java" />
+	</checkstyle>
+</suppressions>
