+++ modules/util/portal-tools-db-partition-migration-validator/src/main/java/com/liferay/portal/tools/db/partition/migration/validator/common/Company.java
@@ -1,0 +1,93 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.portal.tools.db.partition.migration.validator.common;
+
+import java.util.Objects;
+
+/**
+ * @author Luis Ortiz
+ */
+public class Company {
+
+	public Company() {
+	}
+
+	public Company(
+		long companyId, String companyName, String virtualHost, String webId) {
+
+		_companyId = companyId;
+		_companyName = companyName;
+		_webId = webId;
+
+		_virtualHostName = virtualHost;
+	}
+
+	@Override
+	public boolean equals(Object object) {
+		if (this == object) {
+			return true;
+		}
+
+		if (!(object instanceof Company)) {
+			return false;
+		}
+
+		Company company = (Company)object;
+
+		if ((_companyId == company._companyId) &&
+			_companyName.equals(company._companyName) &&
+			_virtualHostName.equals(company._virtualHostName) &&
+			_webId.equals(company._webId)) {
+
+			return true;
+		}
+
+		return false;
+	}
+
+	public long getCompanyId() {
+		return _companyId;
+	}
+
+	public String getCompanyName() {
+		return _companyName;
+	}
+
+	public String getVirtualHostName() {
+		return _virtualHostName;
+	}
+
+	public String getWebId() {
+		return _webId;
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(_companyId, _companyName, _virtualHostName, _webId);
+	}
+
+	public void setCompanyId(long companyId) {
+		_companyId = companyId;
+	}
+
+	public void setCompanyName(String companyName) {
+		_companyName = companyName;
+	}
+
+	public void setVirtualHostName(String virtualHostName) {
+		_virtualHostName = virtualHostName;
+	}
+
+	public void setWebId(String webId) {
+		_webId = webId;
+	}
+
+	private long _companyId;
+	private String _companyName;
+	private String _virtualHostName;
+	private String _webId;
+
+}
