+++ modules/util/portal-tools-db-partition-migration-validator/src/main/resources/com/liferay/portal/tools/db/partition/migration/validator/packageinfo
