+++ modules/util/portal-tools-db-partition-migration-validator/src/test/resources/com/liferay/portal/tools/db/partition/migration/validator/dependencies/targetSuccess.data
