--- modules/util/portal-tools-db-partition-virtual-instance-migration-test/bnd.bnd
@@ -1,15 +1,0 @@
-Bundle-Name: Liferay Portal Tools DB Partition Virtual Instance Migration Test
-Bundle-SymbolicName: com.liferay.portal.tools.db.partition.virtual.instance.migration.test
-Bundle-Version: 1.0.0
-Import-Package:\
-	!com.google.protobuf.*,\
-	\
-	!com.mchange.v2.c3p0.*,\
-	\
-	!com.oracle.bmc.*,\
-	\
-	*
--includeresource:\
-	@com.liferay.portal.tools.db.partition.virtual.instance.migration-*.jar!/com/liferay/portal/tools/db/partition/virtual/instance/migration/*.class,\
-	@com.liferay.portal.tools.db.partition.virtual.instance.migration-*.jar!/org/apache/commons/cli/*.class
--removeheaders: Require-Capability
