--- modules/util/portal-tools-db-partition-virtual-instance-migration-test/build.gradle
@@ -1,7 +1,0 @@
-dependencies {
-	testIntegrationImplementation group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
-	testIntegrationImplementation group: "commons-io", name: "commons-io", version: "2.11.0"
-	testIntegrationImplementation project(":apps:portal:portal-db-partition-test-util")
-	testIntegrationImplementation project(":test:arquillian-extension-junit-bridge")
-	testIntegrationImplementation project(":util:portal-tools-db-partition-virtual-instance-migration")
-}
