--- modules/util/portal-tools-db-partition-virtual-instance-migration/README.markdown
@@ -1,36 +1,0 @@
-# Database Partition Virtual Instance Migration Tool
-This tool allows the export and validation of the database information needed to validate that all the conditions for a successful virtual instance migration are met.
-
-## Requirements
-- MySQL
-- Database user with DDL privileges
-
-## Usage
-```
-./db_partition_virtual_instance_migration.sh -e [OPERATION_MODE] [OPERATION_PARAMETERS]
-```
-
-Main Operation mode:
-
-- `-e`, `--export` Execute data export.
-- `-h`, `--help` Display options.
-- `-v`, `--validate` Execute data validation.
-
-Data Export parameters:
-
-- `-d`, `--output-dir <arg>` Output directory. Folder ./exports otherwise.
-- `-j`, `--jdbc-url <arg>` JDBC URL.
-- `-p`, `--password <arg>` Database user password.
-- `-s`, `--schema-name <arg>` Database schema name to be exported. Default schema name otherwise.
-- `-u`, `--user <arg>` Database user name.
-
-Data Validation parameters:
-
-- `-s`, `--source-file <arg>` Source file generated by the Database Partition Virtual Instance Migration Tool.
-- `-t` , `--target-file <arg>` Target file generated by the Database Partition Virtual Instance Migration Tool.
-
-### Execution examples
-```
-./db_partition_virtual_instance_migration.sh -e -j "jdbc:mysql://localhost:3306/lportal?useUnicode=true&characterEncoding=UTF-8&useFastDateParsing=false&useTimezone=true&serverTimezone=GMT" -u user -p password -s lpartition_1234 -d "/migrations/exports"
-./db_partition_virtual_instance_migration.sh -v -s "/bundle/tools/portal-tools-db-partition-virtual-instance-migration/exports/source.json" -t "/bundle/tools/portal-tools-db-partition-virtual-instance-migration/exports/target.json"
-```
